[VFS/STFS] Move x360-specific STFS structs to stfs_xbox.h
